Restructures the project, once again
