+++ src/core/debug/log.cpp
@@ -1,0 +1,1 @@
+#include "log.hpp"
