+++ src/core/debug/log.hpp
@@ -1,47 +1,21 @@
-/**
- * MIT License
- *
- * Copyright (c) 2016 Rafael C. Nunes
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to
- * deal in the Software without restriction, including without limitation the
- * rights to use, copy, modify, merge, publish, distribute, sublicense, and/or
- * sell copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
- * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
- * IN THE  SOFTWARE.
- */
-
-#ifndef DEBUG_HPP
-#define DEBUG_HPP
+#pragma once
 
 #include <iostream>
 
-#include "macros.hpp"
+#include "colors.hpp"
 
 /**
  * @brief The Debug class is where debug functions are defined, using the power
  * of variadic templates it .
  */
-class Debug {
+class Log {
 public:
     /**
      * @brief log The default log, it prints args on the stdout.
      * @param args The arguments to be printed on the terminal.
      */
     template<typename... Ts>
-    static void log(const Ts&... args)
+    static void info(const Ts&... args)
     {
         std::cout << SUCCESS_COLOR << "LOG: ";
 
@@ -57,7 +31,7 @@ public:
      * @param args The arguments to be printed on the terminal.
      */
     template<typename... Ts>
-    static void log_warn(const Ts&... args)
+    static void warn(const Ts&... args)
     {
         std::cout << WARNING_COLOR << "WARN: ";
 
@@ -73,7 +47,7 @@ public:
      * @param args The arguments to be printed on the terminal.
      */
     template<typename... Ts>
-    static void log_err(const Ts&... args)
+    static void error(const Ts&... args)
     {
         std::cerr << ERROR_COLOR << "ERROR: ";
