+++ src/renderer/renderer.cpp
@@ -1,0 +1,8 @@
+#include "kkr.hpp"
+#include "renderer.hpp"
+
+Renderer::Renderer(Window &&window) {}
+
+Renderer::~Renderer() {}
+
+void Renderer::information() {}
